requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
chromadb>=0.5.0
openai>=1.0.0
sentence-transformers>=3.0.0
ragas>=0.1.0
wandb>=0.17.0
fastapi>=0.111.0
uvicorn>=0.30.0
pypdf>=4.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
torch>=2.0.0
transformers>=4.40.0